backend/requirements.txt
fastapi==0.104.1
pandas==2.1.3
openpyxl==3.1.2
chromadb==0.4.18
groq==0.4.1
python-dotenv==1.0.0
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
uuid==1.30